requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
aiofiles==24.1.0
websockets==13.0
python-multipart==0.0.9
python-dotenv==1.0.1
TikTokLive>=6.6.5
gTTS==2.5.4
edge-tts==6.1.12
httpx==0.27.0
SQLAlchemy==2.0.36
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
mutagen==1.47.0
psycopg[binary]==3.2.3
